Last commit before first compilation of SubWojInt
Finished implementation of optionalRoundings(). Placed
constraintIsFeasible() inside of checkAllConstraints(), the only place
where it was used. Went through all integer code with a fine-toothed
comb. Changed type signatures  of a number of functions. Found the
potential for an infinite loop in forcedRounding, which has been
eliminated. Renamed struct members pertaining to the IntegerTree queue
to be more clear.  Rearranged functions to be in an order that made more
sense.