SubWojInt lives! - All problems appear to be solved.
Plenty of typos and other compilation-time issues resolved. Wasn't initializing
IntegerTree->additionsFirst. Wasn't correctly adding the first IntegerTreeVertex to
the IntegerTree queue. Wasn't creating an empty EdgeRefList to be referenced by the
x0 root node of the IntegerTree.

Leaving a bunch of commented-out print statements that were being used for
debugging, for now. Will probably be gone for the next commit.