Profiling subWojInt
Removed commented-out print statements and such. Added another function to add an edge to the
beginning of an EdgeRefList, so now there is the option of either the beginning or the end.
These options allow system->infeasibilityProof to be ordered such that assumptions are
immediately followed by their associated contradictions, in proofs of integral infeasibility.
Added code to pass profiling information to auto.sh. Created profile_subWojInt.sh to run
auto.sh four times, concurrently. At least one of the input files has uncovered a memory
issue, so that will need to be addressed.