Address compile and run-time errors in last commit
Missed a couple of *[DOWN] *[UP] things. Forgot to regenerate the IntegerTree
after freeing it in the middle of optionalRoundings().

Besides version control, I'm basically using Git in the place of WinSCP to move
things back and forth off of the server I'm using to compile, test, and profile.

Also, in the middle of profiling the program again.