Began lahiriMine
Everything is implemented, up to dijkstra in johnsonAllPairs. I need to
implement the Fibonacci Heap in order to run dijkstra as fast as
possible, so I'm reading over how that works, right now.

I added inttest8, based off of an example in the Lahiri paper.

My computer scared me a bit, so I'm basically using this as a backup, as
well.