Reduced memory usage of lahiri
Using an anonymous union and anonymous structs to decrease the overall size
of the Vertex struct. Two sets of struct members would not be used at the
same time as each other. The set that is used first is still initialized in
initializeSystem(). A function named setSystemForJohnson() initializes the
second set before it's used.