Modify subWojInt and lahiri
For more accurate profiling on WVU's HPC system. lahiri definitely
requires a very large amount of system memory to generate integral
solutions for the larger input systems. subWojInt will have to be run on
the same system and in the same way, in order to make an
apples-to-apples comparison.

This is set up with preprocessor directives, so it can still be compiled
with only C standard library functionality.