Add test/intertest6.txt
This will test for a parseError() failure mode, wherein the line with an
error in it ends with the beginning of a block quote. Currently,
parseError() won't do anything differently with /* than it would with
anything else, and this should lead to getToken() trying to tokenize the
middle of a block quote.