Create constraint.h
Separating definition of Constraint from UTVPIinterpreter, as other
functions may be written to define constraints and feed those into an
inequality system solver.

Also added CONSTRAINT_ to the beginning of ConstraintSign enum members,
so it would be easier to use the enum in other solvers, where PLUS and
MINUS may be used for something else.

Therefore had to modify all the inequality system solver files.