Modify utvpiInterpreter
Move everything that shouldn't be in the header file into the source
file. I'm curious if C can tolerate me having the struct Parser
definition in the source file, as long as typedef struct Parser Parser
is in the header file. It isn't used anywhere else and is just passed
into initializeSystem and addEdge so it can be passed to parseError,
which itself is defined in utvpiInterpreter.