Solve absolute constraint fault
This error has been present since I removed the unnecessary second
IntegerTree from System, more than a month ago. Absolute constraints
were stored in IntegerTree. Freeing and then regenerating the
IntegerTree in optionalRoundings() removes the first absolute constraint
and the second one is written in the exact same place as the first one
was. Thus, the second absolute constraint shows up for both first and
second absolute constraints in the proof of integral infeasibility.

additionsFirst has been moved from IntegerTree to System. A new
function, freeAdditions(), has been added, such that it can be called at
the bottom of the optionalRoundings() loop, and in freeSystem().