delay.h and integration into lahiriCO1
delay.h allows each graph representation to be constructed after
completion of input file parsing, which takes a lot of time and really
ought to be irrelevant. This way, the time that graph representation
construction takes can be integrated into the linear algorithm time
without making that data point useless.